# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Destination {
  id: number;
  name: string;
  duration: string;
  /** Difficulty score 0-100 */
  difficulty: number;
  maxAltitude: string;
  distance: string;
  route: string;
  region: string;
  seatsTotal: number;
  seatsLeft: number;
  bookingStatus: string;
  lat: number;
  lng: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  hotelName?: string | null;
  /** @nullable */
  hotelPhone?: string | null;
  /** @nullable */
  hospitalName?: string | null;
  /** @nullable */
  emergencyPhone?: string | null;
  itinerary?: string[];
  elevationPoints?: number[];
  createdAt?: string;
}

export interface DestinationInput {
  name: string;
  duration: string;
  difficulty: number;
  maxAltitude: string;
  distance: string;
  route: string;
  region: string;
  seatsTotal: number;
  lat: number;
  lng: number;
  imageUrl?: string;
  hotelName?: string;
  hotelPhone?: string;
  hospitalName?: string;
  emergencyPhone?: string;
  itinerary?: string[];
  elevationPoints?: number[];
}

export interface DestinationUpdate {
  name?: string;
  duration?: string;
  difficulty?: number;
  maxAltitude?: string;
  distance?: string;
  route?: string;
  region?: string;
  seatsTotal?: number;
  seatsLeft?: number;
  bookingStatus?: string;
  lat?: number;
  lng?: number;
  imageUrl?: string;
  hotelName?: string;
  hotelPhone?: string;
  hospitalName?: string;
  emergencyPhone?: string;
  itinerary?: string[];
  elevationPoints?: number[];
}

export interface Trek {
  id: number;
  userId: number;
  destinationId: number;
  /** @nullable */
  destinationName?: string | null;
  /** planned | active | completed | cancelled */
  status: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  groupSize: number;
  /** @nullable */
  guideId?: number | null;
  /** @nullable */
  guideName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completionRating?: number | null;
  createdAt: string;
}

export interface TrekInput {
  userId: number;
  destinationId: number;
  startDate: string;
  groupSize: number;
  /** @nullable */
  guideId?: number | null;
  notes?: string;
}

export interface TrekUpdate {
  status?: string;
  endDate?: string;
  /** @nullable */
  guideId?: number | null;
  notes?: string;
  completionRating?: number;
}

export interface Guide {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  experience: string;
  languages: string;
  rating: number;
  license: string;
  available: boolean;
  /** @nullable */
  destinationId?: number | null;
  totalTreks?: number;
  createdAt?: string;
}

export interface GuideInput {
  name: string;
  phone?: string;
  experience: string;
  languages: string;
  rating: number;
  license: string;
  available?: boolean;
  /** @nullable */
  destinationId?: number | null;
}

export interface GuideUpdate {
  name?: string;
  phone?: string;
  experience?: string;
  languages?: string;
  rating?: number;
  license?: string;
  available?: boolean;
  /** @nullable */
  destinationId?: number | null;
}

export interface User {
  id: number;
  username: string;
  /** user | admin */
  role: string;
  totalTreks?: number;
  completedTreks?: number;
  createdAt: string;
}

export interface UserInput {
  username: string;
  password: string;
  role?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserUpdate {
  username?: string;
  password?: string;
}

export interface AnalyticsSummary {
  totalUsers: number;
  totalTreks: number;
  activeBookings: number;
  totalDestinations: number;
  completedTreks: number;
  totalGuides: number;
}

export interface PopularRoute {
  destinationId: number;
  destinationName: string;
  bookingCount: number;
  difficulty: number;
}

export interface MonthlyBooking {
  month: string;
  count: number;
}

export interface DifficultyBreakdown {
  level: string;
  count: number;
}

export interface WishlistItem {
  id: number;
  userId: number;
  destinationId: number;
  createdAt: string;
  destination?: Destination;
}

export interface WishlistInput {
  userId: number;
  destinationId: number;
}

export interface Review {
  id: number;
  userId: number;
  destinationId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  username: string;
  createdAt: string;
}

export interface ReviewInput {
  userId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
}

export type ListDestinationsParams = {
difficulty?: string;
duration?: string;
region?: string;
};

export type ListTreksParams = {
userId?: number;
status?: string;
};

export type ListGuidesParams = {
available?: boolean;
destinationId?: number;
};

export type ListWishlistParams = {
userId: number;
};