# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  name: string;
  duration: string;
  /** Difficulty score 0-100 */
  difficulty: number;
  maxAltitude: string;
  distance: string;
  route: string;
  region: string;
  seatsTotal: number;
  seatsLeft: number;
  bookingStatus: string;
  lat: number;
  lng: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  hotelName?: string | null;
  /** @nullable */
  hotelPhone?: string | null;
  /** @nullable */
  hospitalName?: string | null;
  /** @nullable */
  emergencyPhone?: string | null;
  itinerary?: string[];
  elevationPoints?: number[];
  createdAt?: string;
}