# lib/api-zod/src/generated/types/destinationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface DestinationInput {
  name: string;
  duration: string;
  difficulty: number;
  maxAltitude: string;
  distance: string;
  route: string;
  region: string;
  seatsTotal: number;
  lat: number;
  lng: number;
  imageUrl?: string;
  hotelName?: string;
  hotelPhone?: string;
  hospitalName?: string;
  emergencyPhone?: string;
  itinerary?: string[];
  elevationPoints?: number[];
}