# lib/api-zod/src/generated/types/trek.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface Trek {
  id: number;
  userId: number;
  destinationId: number;
  /** @nullable */
  destinationName?: string | null;
  /** planned | active | completed | cancelled */
  status: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  groupSize: number;
  /** @nullable */
  guideId?: number | null;
  /** @nullable */
  guideName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completionRating?: number | null;
  createdAt: string;
}