# lib/api-zod/src/generated/types/trekInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface TrekInput {
  userId: number;
  destinationId: number;
  startDate: string;
  groupSize: number;
  /** @nullable */
  guideId?: number | null;
  notes?: string;
}