# lib/api-zod/src/generated/types/trekUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trail Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface TrekUpdate {
  status?: string;
  endDate?: string;
  /** @nullable */
  guideId?: number | null;
  notes?: string;
  completionRating?: number;
}